Module OManifold

Constructive Solid Geometry

module Cross : sig ... end

Two-dimensional cross sections guaranteed to be without self-intersections, or overlaps between polygons (from construction onwards)

module Manifold : sig ... end

Manifold triangular meshes representing 3d solid objects

module Quality : sig ... end

Quality globals akin to OpenSCAD's facet governing "special" parameters, $fn, $fa, and $fs. These are global variables that are handy for quickly switching from quick and dirty rough iteration quality to computationally expensive smooth final products.

Bounds

module MRect : sig ... end

The 2d axis-aligned recantagle (bounding box) representation of the Manifold library

module MBox : sig ... end

The 3d axis-aligned bounding box representation of the Manifold library

Signed Distance Functions

module Sdf2 : sig ... end

module Sdf3 : sig ... end

IO

module MMeshGL : sig ... end

A graphics library friendly representation of manifold's internal mesh. Obtained via Manifold.to_mmeshgl, or constructed directly with MMeshGL.make

module Export : sig ... end

Writing manifold meshes to disk via Open Asset Import Library (assimp)