Module OManifold.Cross

Two-dimensional cross sections guaranteed to be without self-intersections, or overlaps between polygons (from construction onwards)

type t

type fill_rule = [

| `EvenOdd
(*
only odd numbered sub-regions are filled
*)

| `NonZero
(*
non-zero sub-regions are filled
*)

| `Positive
(*
only sub-regions with winding counts > 0 are filled
*)

| `Negative
(*
only sub-regions with winding counts < 0 are filled
*)

]

Filling rules used by the clipping algorithm for boolean operations. See Clipper2's docs for a detailed explanation of how they differ).

Constructors

val empty : unit -> t

empty ()

Create an empty cross-section containing no contours.

val of_path : ?fill_rule:fill_rule -> OCADml.Path2.t -> t

of_path ?fill_rule path

Create a 2d cross-section from a single outline path. A boolean union operation (with `Positive filling rule by default) is performed to ensure the resulting cross-section is free of self-intersections.

val of_paths : ?fill_rule:fill_rule -> OCADml.Path2.t list -> t

of_paths ?fill_rule paths

Create a 2d cross-section from a set of closed paths (zero or more complex polygons). A boolean union operation (with `Positive filling rule by default) is performed to combine overlapping polygons and ensure the resulting cross-section is free of intersections.

val of_poly2 : ?fill_rule:fill_rule -> OCADml.Poly2.t -> t

of_poly2 ?fill_rule poly

Create a 2d cross-section from an OCADml polygon. A boolean union operation (with `Positive filling rule by default) is performed to combine overlapping polygons and ensure the resulting cross-section is free of intersections.

val of_poly2s : ?fill_rule:fill_rule -> OCADml.Poly2.t list -> t

of_poly2 ?fill_rule poly

Create a 2d cross-section from a list of OCADml polygons. A boolean union operation (with `Positive filling rule by default) is performed to combine overlapping polygons and ensure the resulting cross-section is free of intersections.

Shapes

val circle : ?fn:int -> float -> t

circle ?fn r

Create a circular cross-section with radius r. fn can be provided to set the number of segments that it should be drawn with, otherwise it will be determined according to the globals governed by OManifold.Quality.

val square : ?center:bool -> OCADml.v2 -> t

square ?center dims

Create a square with the given XY dims. By default it is positioned in the first quadrant, touching the origin. Setting ~center:true will instead centre the shape on the origin. If any dimensions in size are negative, or if all are zero, an empty Manifold.t will be returned.

Booleans

val boolean : op:[ `Add | `Subtract | `Intersect ] -> t -> t -> t

boolean ~op a b

Compute the boolean operation op between the cross-sections a and b.

val batch_boolean : op:[ `Add | `Subtract | `Intersect ] -> t list -> t

batch_boolean ~op cs

Compute the boolean operation op on the list of cross-sections cs. In the case of ~op:`Subtract, the tail is differenced from the head.

val add : t -> t -> t

add a b

Union (logical or) the cross-sections a and b.

val union : t list -> t

union ts

Union (logical or) the list of cross-sections ts.

val sub : t -> t -> t

sub a b

Subtract (logical and not) the cross-section b from the cross-section a.

val difference : t -> t list -> t

difference t d

Subtract (logical and not) the list of cross-sections d from the cross-section t.

val intersect : t -> t -> t

intersect a b

Compute the intersection (logical and) of the cross-sections a and b.

val intersection : t list -> t

intersection ts

Compute the intersection (logical and) of the cross-sections ts. Only the area which is common or shared by all shapes are retained. If ts is empty, an empty cross-section t will result.

val rect_clip : t -> OCADml.Box2.t -> t

rect_clip t rect

Compute the intersection between the cross-section t and an axis-aligned bounding box rect. This operation has much higher performance (O(n) vs O(n3) than the general purpose intersection algorithm used for sets of cross-sections.

Topology

val compose : t list -> t

compose ts

Create a cross-section from a list of cross-sections ts (see union). Unlike Manifold.compose, it isn't necessary that the input shapes are non-overlapping as a clipping operation is performed here.

val decompose : t -> t list

decompose t

Break the cross-section t into a list of topologically disconnected cross-sections, each containing one outline contour with zero or more holes.

Transformations

val translate : OCADml.v2 -> t -> t

val xtrans : float -> t -> t

val ytrans : float -> t -> t

val rotate : ?about:OCADml.v2 -> float -> t -> t

val zrot : ?about:OCADml.v2 -> float -> t -> t

val mirror : OCADml.v2 -> t -> t

val affine : OCADml.Affine2.t -> t -> t

val scale : OCADml.v2 -> t -> t

val xscale : float -> t -> t

val yscale : float -> t -> t

val warp : (OCADml.v2 -> OCADml.v2) -> t -> t

warp f t

Map over the vertices of the cross-section t with the function f, followed by a union operation ensuring that the result is free of intersections.

Path simplification and offsetting

val simplify : ?eps:float -> t -> t

val offset : ?join_type:[< `Miter | `Round | `Square Square ] -> ?miter_limit:float -> ?arc_tolerance:float -> delta:float -> t -> t

Geometry

val bounds : t -> OCADml.Box2.t

val area : t -> float

val num_vert : t -> int

val num_contour : t -> int

val is_empty : t -> bool

Conversion

val to_paths : t -> OCADml.Path2.t list