Module OManifold.MBox

type t

The 3-dimensional bounding box representation of the Manifold library.

Constructor

val make : OCADml.v3 -> OCADml.v3 -> t

make a b

Create a bounding box that contains the points a and b.

Data / Queries

val min : t -> OCADml.v3

min t

The minimum bound of the box t.

val max : t -> OCADml.v3

max t

The maximum bound of the box t.

val center : t -> OCADml.v3

center t

The center point of the box t.

val dimensions : t -> OCADml.v3

dimensions t

The 3-dimensional extent (xyz) of the box t.

val abs_max_coord : t -> float

abs_max_coord t

The absolute largest dimensional coordinate of the box t.

val contains_pt : t -> OCADml.v3 -> bool

contains_pt t p

Determine whether the box t contains the point p.

val contains_box : t -> t -> bool

contains_box a b

Determine whether the box a contains the box b.

val overlaps_pt : t -> OCADml.v3 -> bool

overlaps_pt t p

Determine whether the projection/shadow of the box t on the xy plane contains that of the point p.

val overlaps_box : t -> t -> bool

overlaps_box a b

Determine whether the projection/shadow of the box a on the xy plane overlaps that of the box b.

val is_finite : t -> bool

is_finite t

Determine whether the box t is finite.

Booleans

val union : t -> t -> t

union a b

Return the union of the boxes a and b (enveloping the space of both).

val include_pt : t -> OCADml.v3 -> unit

include_pt t p

Expand the box t (mutating it) to include the point p.

Transformations

val translate : OCADml.v3 -> t -> t

translate p t

Move the box t along the vector p.

val scale : OCADml.v3 -> t -> t

scale factors t

Scale (multiply) the box t by the given xyz factors.

val transform : OCADml.Affine3.t -> t -> t

transform m t

Transform box t with the affine transformation matrix m.

OCADml conversion

val of_box : OCADml.Box3.t -> t

of_box b

Construct a manifold box from the bounds of the 3d OCADml bounding box b.

val to_box : t -> OCADml.Box3.t

to_box t

Return the minimum and maximum bounds of the manifold box t as an OCADml bounding box.