Module OManifold.MMesh

type properties = {

tris : (int * int * int) list;
(*
Indices into props for each tri of the mesh
*)

props : float list;
(*
Sets of property values, the length of which should be the largest index in tris times the length of tolerances (the number of properties). Think of it as a property matrix indexed as idx * n_tolerances + tolerance_idx.
*)

tolerances : float list;
(*
Precision values. This is the amount of interpolation error allowed before two neighouring triangles are considered not coplanar. A good place to start is 1e-5 times the largest value you expect this property to take.
*)

}

Mesh properties which can optionally be provided to Manifold.of_mmesh.

type t

The 3-dimensional mesh representation of the Manifold library.

Constructor

val make : ?normals:OCADml.v3 list -> ?tangents:OCADml.v4 list -> OCADml.v3 list -> (int * int * int) list -> t

make ?normals ?tangents points triangles

Create a manifold mesh from a set of points, and triangles represented as triples of indices into points. Vertex normals and half-edge tangents can optionally be provided. If so, they must be the same length as points and 3x the length of triangles respectively -- Invalid_argument will be raised otherwise.

Data Extraction

val points : t -> OCADml.v3 list

points t

Retrieve the points making up the mesh t.

val triangles : t -> (int * int * int) list

triangles t

Retrieve the triangular faces of the mesh t as triples of indices into its points.

val normals : t -> OCADml.v3 list

normals t

Retrieve the vertex normals of the mesh t (same length as points).

val halfedge_tangents : t -> OCADml.v4 list

halfedge_tangents t

Retrieve the halfedge tangents of the mesh t (three for each of the t's triangles).

OCADml conversion

To and from the OCADml Mesh.t type. As Mesh.t follows the OpenSCAD winding convention (opposite to manifold) the triangles/faces are reversed.

val of_mesh : OCADml.Mesh.t -> t

of_mesh m

Create a manifold mesh from the OCADml.Mesh.t m.

val to_mesh : t -> OCADml.Mesh.t

to_mesh t

Create an OCADml.Mesh.t from the manifold mesh t.