Module Manifold.Curvature

module Bounds : sig ... end

type t = {

bounds : Bounds.t;

vert_mean : float list;

vert_gaussian : float list;

}

Computed vertex curvatures of a Manifold.t

The inverse of the radius of curvature, and signed such that positive is convex and negative is concave. There are two orthogonal principal curvatures at any point on a manifold, with one maximum and the other minimum. Gaussian curvature is their product, while mean curvature is their sum. This approximates them for every vertex (returned as vectors in the structure) and also returns their minimum and maximum values.