Module MeshRelation.BaryRef

type t = {

mesh_id : int;
(*
unique ID to the particular instance of the mesh
*)

original_id : int;
(*
ID corresponding to the source mesh
*)

tri : int;
(*
triangle index in the source mesh to which vert_bary refers
*)

vert_bary : int * int * int;
(*
an index for each vertex into the barycentric list if that index is >= 0, indicating it is a new vertex, otherwise, if the index is < 0, this indicates it is an original vertex (the index + 3 vertex of the reference triangle)
*)

}