Module OManifold.Polygons

type t

Constructors

val make : OCADml.Path2.t list -> t

make paths

Create a set of polygons from a set of 2d paths. These can vary from arbitrarily nested to completely separate, describing one or more complex polygons with holes.

OCADml Conversion

These are nothing special, but are provided for convenience (prepending the outer path onto the list of holes before applying make).

val of_path : OCADml.Path2.t -> t

of_path path

Lift a 2d OCADml.Path2.t into a manifold polygon set.

val of_paths : OCADml.Path2.t list -> t

of_paths paths

Lump together a set of 2d OCADml.Path2.ts into a manifold polygon set. Equivalent to make.

val of_poly2 : OCADml.Poly2.t -> t

of_poly2 poly

Lift a 2d OCADml.Poly2.t into a manifold polygon set.

val of_poly2s : OCADml.Poly2.t list -> t

of_poly2s polys

Lump together a set of 2d OCADml.Poly2.ts into a manifold polygon set.