Module OManifold

type mrect

type cross_section

module MMeshGL : sig ... end

type op_type = [

| `Add

| `Subtract

| `Intersect

]

module Cross : sig ... end

module MRect : sig ... end

module MBox : sig ... end

module Manifold : sig ... end

module Sdf2 : sig ... end

module Sdf3 : sig ... end

module Quality : sig ... end

Quality globals akin to OpenSCAD's facet governing "special" parameters, $fn, $fa, and $fs. These are global variables that are handy for quickly switching from quick and dirty rough iteration quality to computationally expensive smooth final products.

module Export : sig ... end

Writing manifold meshes to disk via Open Asset Import Library (assimp).