Module OManifold.Cross

type t = cross_section

type fill_rule = [

| `EvenOdd

| `Negative

| `NonZero

| `Positive

]

Constructors

val empty : unit -> t

val copy : t -> t

val of_path : ?fill_rule:fill_rule -> OCADml.Path2.t -> t

val of_paths : ?fill_rule:fill_rule -> OCADml.Path2.t list -> t

val of_poly2 : ?fill_rule:fill_rule -> OCADml.Poly2.t -> t

val of_poly2s : ?fill_rule:fill_rule -> OCADml.Poly2.t list -> t

Shapes

val circle : ?fn:int -> float -> t

val square : ?center:bool -> OCADml.v2 -> t

Booleans

val boolean : op:op_type -> t -> t -> t

val batch_boolean : op:op_type -> t list -> t

val add : t -> t -> t

val sub : t -> t -> t

val intersect : t -> t -> t

val union : t list -> t

val difference : t -> t list -> t

val intersection : t list -> t

val rect_clip : t -> mrect -> t

Topology

val compose : t list -> t

val decompose : t -> t list

Transformations

val translate : OCADml.v2 -> t -> t

val xtrans : float -> t -> t

val ytrans : float -> t -> t

val rotate : ?about:OCADml.v2 -> float -> t -> t

val zrot : ?about:OCADml.v2 -> float -> t -> t

val mirror : OCADml.v2 -> t -> t

val affine : OCADml.Affine2.t -> t -> t

val scale : OCADml.v2 -> t -> t

val xscale : float -> t -> t

val yscale : float -> t -> t

val warp : (OCADml.v2 -> OCADml.v2) -> t -> t

warp f t

Map over the vertices of the cross-section t with the function f, followed by a union operation ensuring that the result is free of intersections.

Path simplification and offsetting

val simplify : ?eps:float -> t -> t

val offset : ?join_type:[< `Miter | `Round | `Square Square ] -> ?miter_limit:float -> ?arc_tolerance:float -> delta:float -> t -> t

Geometry

val bounds : t -> mrect

val area : t -> float

val num_vert : t -> int

val num_contour : t -> int

val is_empty : t -> bool

Conversion

val to_paths : t -> OCADml.Path2.t list